refactor: Restyle navbar and restructure website layout
The navbar has been restyled to look considerably better than before,
with a fully responsive mobile view. Space for the TaskCollect has been
added, although currently it only has the text. A logo will replace
this in a future commit.

The website layout has also been restructured in an effort to not only
make developing the website easier, but also out of necessity for
certain features (e.g. scrollbar that stays below the navbar) to work.